Fix warning: implicit conversion changes signedness
visp/tutorial/image/tutorial-canny-hsv.cpp:309:63: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  309 |   vpImageFilter::gaussianBlur(Iin_hsvuc, Iblur_hsvuc, options.m_gaussianKernelSize, options.m_gaussianStdev, true, p_mask);
      |   ~~~~~~~~~~~~~                                       ~~~~~~~~^~~~~~~~~~~~~~~~~~~~
visp/tutorial/image/tutorial-canny-hsv.cpp:321:61: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  321 |   vpImageFilter::gaussianBlur(Iin_hsvd, Iblur_hsvd, options.m_gaussianKernelSize, options.m_gaussianStdev, true, p_mask);
      |   ~~~~~~~~~~~~~                                     ~~~~~~~~^~~~~~~~~~~~~~~~~~~~
visp/tutorial/image/tutorial-canny-hsv.cpp:331:61: warning: implicit conversion changes signedness: 'int' to 'unsigned int' [-Wsign-conversion]
  331 |   vpImageFilter::gaussianBlur(Iin_convert, IblurUC, options.m_gaussianKernelSize, (FilterType)options.m_gaussianStdev, true, p_mask);
      |   ~~~~~~~~~~~~~                                     ~~~~~~~~^~~~~~~~~~~~~~~~~~~~
visp/tutorial/image/tutorial-canny-hsv.cpp:382:39: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  382 |     int posX = disp_input->getWidth() + 20;
      |         ~~~~   ~~~~~~~~~~~~~~~~~~~~~~~^~~~
visp/tutorial/image/tutorial-canny-hsv.cpp:383:40: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  383 |     int posY = disp_input->getHeight() + 20;
      |         ~~~~   ~~~~~~~~~~~~~~~~~~~~~~~~^~~~

+++ modules/core/include/visp3/core/vpImageFilter.h
@@ -473,7 +473,7 @@ public:
   {
     const unsigned int w = I.getWidth();
     const unsigned int h = I.getHeight();
-    const int size = I.getSize();
+    const int size = static_cast<int>(I.getSize());
 
     if ((lowerThresholdRatio <= 0.f) || (lowerThresholdRatio >= 1.f)) {
       std::stringstream errMsg;
@@ -566,7 +566,7 @@ public:
     vpHistogram hist(nbBins);
     hist.setMask(p_mask);
     OutType step = 0.;
-    hist.calculate<OutType>(dI, dIMin, dIMax, step, nbBins, nbThread);
+    hist.calculate<OutType>(dI, dIMin, dIMax, step, nbBins, static_cast<unsigned int>(nbThread));
     float totalNbPixels = static_cast<float>(hist.getTotal());
     float accu = 0;
     float t = upperThresholdRatio * totalNbPixels;
@@ -1549,7 +1549,7 @@ public:
 #ifdef VISP_HAVE_OPENMP
       }
 #endif
-      }
+    }
   }
 
   /**
